+++ base.php
@@ -1,8 +1,10 @@
 <?php
 
 namespace Roots\Sage;
+
 use Roots\Sage\Config;
 use Roots\Sage\Wrapper;
+
 ?>
 
 <?php get_template_part('templates/head'); ?>